// src/api/models/models-WaypointTraitSymbol/waypointTraitSymbol.ts
/**
 * Generated by orval v8.0.0-rc.0 🍺
 * Do not edit manually.
 * SpaceTraders API
 * SpaceTraders is an open-universe game and learning platform that offers a set of HTTP endpoints to control a fleet of ships and explore a multiplayer universe.

The API is documented using [OpenAPI](https://github.com/SpaceTradersAPI/api-docs). You can send your first request right here in your browser to check the status of the game server.

```json http
{
  "method": "GET",
  "url": "https://api.spacetraders.io/v2",
}
```

Unlike a traditional game, SpaceTraders does not have a first-party client or app to play the game. Instead, you can use the API to build your own client, write a script to automate your ships, or try an app built by the community.

We have a [Discord channel](https://discord.com/invite/jh6zurdWk5) where you can share your projects, ask questions, and get help from other players.



 * OpenAPI spec version: 2.3.0
 */

/**
 * The unique identifier of the trait.
 */
export type WaypointTraitSymbol = typeof WaypointTraitSymbol[keyof typeof WaypointTraitSymbol];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WaypointTraitSymbol = {
  UNCHARTED: 'UNCHARTED',
  UNDER_CONSTRUCTION: 'UNDER_CONSTRUCTION',
  MARKETPLACE: 'MARKETPLACE',
  SHIPYARD: 'SHIPYARD',
  OUTPOST: 'OUTPOST',
  SCATTERED_SETTLEMENTS: 'SCATTERED_SETTLEMENTS',
  SPRAWLING_CITIES: 'SPRAWLING_CITIES',
  MEGA_STRUCTURES: 'MEGA_STRUCTURES',
  PIRATE_BASE: 'PIRATE_BASE',
  OVERCROWDED: 'OVERCROWDED',
  HIGH_TECH: 'HIGH_TECH',
  CORRUPT: 'CORRUPT',
  BUREAUCRATIC: 'BUREAUCRATIC',
  TRADING_HUB: 'TRADING_HUB',
  INDUSTRIAL: 'INDUSTRIAL',
  BLACK_MARKET: 'BLACK_MARKET',
  RESEARCH_FACILITY: 'RESEARCH_FACILITY',
  MILITARY_BASE: 'MILITARY_BASE',
  SURVEILLANCE_OUTPOST: 'SURVEILLANCE_OUTPOST',
  EXPLORATION_OUTPOST: 'EXPLORATION_OUTPOST',
  MINERAL_DEPOSITS: 'MINERAL_DEPOSITS',
  COMMON_METAL_DEPOSITS: 'COMMON_METAL_DEPOSITS',
  PRECIOUS_METAL_DEPOSITS: 'PRECIOUS_METAL_DEPOSITS',
  RARE_METAL_DEPOSITS: 'RARE_METAL_DEPOSITS',
  METHANE_POOLS: 'METHANE_POOLS',
  ICE_CRYSTALS: 'ICE_CRYSTALS',
  EXPLOSIVE_GASES: 'EXPLOSIVE_GASES',
  STRONG_MAGNETOSPHERE: 'STRONG_MAGNETOSPHERE',
  VIBRANT_AURORAS: 'VIBRANT_AURORAS',
  SALT_FLATS: 'SALT_FLATS',
  CANYONS: 'CANYONS',
  PERPETUAL_DAYLIGHT: 'PERPETUAL_DAYLIGHT',
  PERPETUAL_OVERCAST: 'PERPETUAL_OVERCAST',
  DRY_SEABEDS: 'DRY_SEABEDS',
  MAGMA_SEAS: 'MAGMA_SEAS',
  SUPERVOLCANOES: 'SUPERVOLCANOES',
  ASH_CLOUDS: 'ASH_CLOUDS',
  VAST_RUINS: 'VAST_RUINS',
  MUTATED_FLORA: 'MUTATED_FLORA',
  TERRAFORMED: 'TERRAFORMED',
  EXTREME_TEMPERATURES: 'EXTREME_TEMPERATURES',
  EXTREME_PRESSURE: 'EXTREME_PRESSURE',
  DIVERSE_LIFE: 'DIVERSE_LIFE',
  SCARCE_LIFE: 'SCARCE_LIFE',
  FOSSILS: 'FOSSILS',
  WEAK_GRAVITY: 'WEAK_GRAVITY',
  STRONG_GRAVITY: 'STRONG_GRAVITY',
  CRUSHING_GRAVITY: 'CRUSHING_GRAVITY',
  TOXIC_ATMOSPHERE: 'TOXIC_ATMOSPHERE',
  CORROSIVE_ATMOSPHERE: 'CORROSIVE_ATMOSPHERE',
  BREATHABLE_ATMOSPHERE: 'BREATHABLE_ATMOSPHERE',
  THIN_ATMOSPHERE: 'THIN_ATMOSPHERE',
  JOVIAN: 'JOVIAN',
  ROCKY: 'ROCKY',
  VOLCANIC: 'VOLCANIC',
  FROZEN: 'FROZEN',
  SWAMP: 'SWAMP',
  BARREN: 'BARREN',
  TEMPERATE: 'TEMPERATE',
  JUNGLE: 'JUNGLE',
  OCEAN: 'OCEAN',
  RADIOACTIVE: 'RADIOACTIVE',
  MICRO_GRAVITY_ANOMALIES: 'MICRO_GRAVITY_ANOMALIES',
  DEBRIS_CLUSTER: 'DEBRIS_CLUSTER',
  DEEP_CRATERS: 'DEEP_CRATERS',
  SHALLOW_CRATERS: 'SHALLOW_CRATERS',
  UNSTABLE_COMPOSITION: 'UNSTABLE_COMPOSITION',
  HOLLOWED_INTERIOR: 'HOLLOWED_INTERIOR',
  STRIPPED: 'STRIPPED',
} as const;