// src/api/models/models-WaypointType/waypointType.ts
/**
 * Generated by orval v8.0.0-rc.1 🍺
 * Do not edit manually.
 * SpaceTraders API
 * SpaceTraders is an open-universe game and learning platform that offers a set of HTTP endpoints to control a fleet of ships and explore a multiplayer universe.

The API is documented using [OpenAPI](https://github.com/SpaceTradersAPI/api-docs). You can send your first request right here in your browser to check the status of the game server.

```json http
{
  "method": "GET",
  "url": "https://api.spacetraders.io/v2",
}
```

Unlike a traditional game, SpaceTraders does not have a first-party client or app to play the game. Instead, you can use the API to build your own client, write a script to automate your ships, or try an app built by the community.

We have a [Discord channel](https://discord.com/invite/jh6zurdWk5) where you can share your projects, ask questions, and get help from other players.



 * OpenAPI spec version: 2.3.0
 */

/**
 * The type of waypoint.
 */
export type WaypointType = (typeof WaypointType)[keyof typeof WaypointType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WaypointType = {
  ARTIFICIAL_GRAVITY_WELL: 'ARTIFICIAL_GRAVITY_WELL',
  ASTEROID: 'ASTEROID',
  ASTEROID_BASE: 'ASTEROID_BASE',
  ASTEROID_FIELD: 'ASTEROID_FIELD',
  DEBRIS_FIELD: 'DEBRIS_FIELD',
  ENGINEERED_ASTEROID: 'ENGINEERED_ASTEROID',
  FUEL_STATION: 'FUEL_STATION',
  GAS_GIANT: 'GAS_GIANT',
  GRAVITY_WELL: 'GRAVITY_WELL',
  JUMP_GATE: 'JUMP_GATE',
  MOON: 'MOON',
  NEBULA: 'NEBULA',
  ORBITAL_STATION: 'ORBITAL_STATION',
  PLANET: 'PLANET',
} as const;